+++ night.go
@@ -187,7 +187,7 @@ func drawNightOverlay(screen *ebiten.Image, ox, oy int) {
 	}
 
 	op := &ebiten.DrawImageOptions{}
-	op.GeoM.Scale(float64(gs.Scale), float64(gs.Scale))
+	op.GeoM.Scale(gs.Scale, gs.Scale)
 	alpha := float32(lvl) / 100.0
 	op.ColorScale.ScaleAlpha(alpha)
 	op.GeoM.Translate(float64(ox), float64(oy))
